Add 3 basic functions
We have a function that creates and manages the Tic-Tac-Toe board.
The function is called 'gameboard' and through a for loop it creates
a 2D-Array grid that represents the board of the game.

Inside this function we also have two closures that that allow the
call of the board (getBoard()) and to update the contents of the cells
inside the board (updateCell()).

We also have another standalone function, called playerController() that
has a private variable that represents the current player to move.
Inside this function are two closures that allows the return of the
current player to move, and to change turns.